+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 * Allow override of Sass compiler from Maven POM or Settings.
 * Removed unused Sass and JavaScript.
 * Enable use of a JavaScript Sass compiler ((sass)[https://www.npmjs.com/package/sass]) instead of Vaadin Sass Compiler.
+* Server-side validation of File(s) uploaded using WFileWidget and WMultiFileWidget #1079.
 
 ### Bug Fixes